CORE-2211: Removes scripts and files related to Dbninja
We've wanted to deprecate the use of Dbninja for a while; having our production database so easily accessible via the browser posed a major security risk. Although each user needed two sets of credentials to authenticate, these could easily be shared around. Instead, everyone now access it via SSH using a proper SQL client which is a much safer approach.
We also enjoy the benefit of running less services in our Kubernetes cluster and reducing of our AWS bill, neiher of which are bad things.